// src/app/Flex/Test/models/PMS060_Search_Criteria.ts
// tslint:disable-next-line:class-name
export class PMS060_Search_Criteria {
    REQUEST_DATE_FROM: Date;
    REQUEST_DATE_TO: Date;
    TEST_DATE_FROM: Date;
    TEST_DATE_TO: Date;
    COMPLETE_DATE_FROM: Date;
    COMPLETE_DATE_TO: Date;
    CHECK_REP_NO_FROM: string;
    CHECK_REP_NO_TO: string;
    PONUMBER_FROM: string;
    PONUMBER_TO: string;
    PERSONINCHARGE: string;
    REQUESTER: string;
    MACHINE_LOC_CD: string;
    CUR_PERSON: string;
    MACHINE_NO_FROM: string;
    MACHINE_NO_TO: string;
    MACHINE_NAME: string;
    VENDORID: number;
    SCHEDULE_TYPEID: number;
    STATUSID: string;
}